% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{cauchy}
\alias{cauchy}
\title{Specify a half-Cauchy prior}
\usage{
cauchy(scale)
}
\arguments{
\item{scale}{Scale parameter (must be positive).}
}
\value{
A \code{did_prior} object.
}
\description{
Used for scale parameters. The distribution is implicitly half-Cauchy
(positive support only) when applied to a \verb{<lower=0>} Stan parameter.
}
