% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{gamma}
\alias{gamma}
\title{Specify a gamma prior}
\usage{
gamma(shape, rate)
}
\arguments{
\item{shape}{Shape parameter (must be positive).}

\item{rate}{Rate parameter (must be positive).}
}
\value{
A \code{did_prior} object.
}
\description{
Used for the degrees-of-freedom parameter when \code{robust_heterogeneity = TRUE}.
}
