% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{print.meta_did_fit}
\alias{print.meta_did_fit}
\title{Print a meta_did_fit object}
\usage{
\method{print}{meta_did_fit}(x, prob = 0.9, ...)
}
\arguments{
\item{x}{A \code{meta_did_fit} object.}

\item{prob}{Width of the credible interval. Default \code{0.9}.}

\item{...}{Ignored.}
}
\description{
Displays a brief summary of the population treatment effect posterior.
}
