% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{tidy.meta_did_fit}
\alias{tidy.meta_did_fit}
\title{Tidy a meta_did_fit object}
\usage{
tidy.meta_did_fit(x, ...)
}
\arguments{
\item{x}{A \code{meta_did_fit} object.}

\item{...}{Ignored.}
}
\value{
\code{NULL} (not yet implemented).
}
\description{
Tidy a meta_did_fit object
}
