% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{normal}
\alias{normal}
\title{Specify a normal prior}
\usage{
normal(mean, sd)
}
\arguments{
\item{mean}{Prior mean.}

\item{sd}{Prior standard deviation (must be positive).}
}
\value{
A \code{did_prior} object.
}
\description{
Specify a normal prior
}
